+++ premium/backend/src/baserow_premium/api/row_comments/serializers.py
@@ -1,31 +1,31 @@
 from django.conf import settings
 from django.contrib.auth import get_user_model
+from rest_framework import serializers
 from rest_framework.fields import CharField
-from rest_framework.serializers import ModelSerializer
 
 from baserow_premium.row_comments.models import RowComment
 
 User = get_user_model()
 
 
-class RowCommentSerializer(ModelSerializer):
-    first_name = CharField(max_length=32, source="user.first_name")
+class RowCommentSerializer(serializers.ModelSerializer):
+    first_name = serializers.CharField(max_length=32, source="user.first_name")
 
     class Meta:
         model = RowComment
         fields = (
             "id",
-            "table",
+            "table_id",
             "row_id",
             "comment",
-            "user",
             "first_name",
             "created_on",
             "updated_on",
+            "user_id",
         )
 
 
-class RowCommentCreateSerializer(ModelSerializer):
+class RowCommentCreateSerializer(serializers.ModelSerializer):
     comment = CharField(max_length=settings.MAX_ROW_COMMENT_LENGTH)
 
     class Meta: