+++ web-frontend/modules/database/applicationTypes.js
@@ -3,10 +3,32 @@ import Sidebar from '@baserow/modules/database/components/sidebar/Sidebar'
 import TemplateSidebar from '@baserow/modules/database/components/sidebar/TemplateSidebar'
 import TableTemplate from '@baserow/modules/database/components/table/TableTemplate'
 import { populateTable } from '@baserow/modules/database/store/table'
+import GridViewRowExpandButton from '@baserow/modules/database/components/view/grid/GridViewRowExpandButton'
 
 export class DatabaseApplicationType extends ApplicationType {
   static getType() {
     return 'database'
+  }
+
+  /**
+   * By default there is no right sidebar in the row edit modal. Override this method
+   * and provide a Sidebar component class if you wish there to be one. This component
+   * will be provided two props row and table.
+   *
+   * @return The component to use as the row edit modal's right sidebar or null to not
+   *    use one.
+   */
+  getRowEditModalRightSidebarComponent() {
+    return null
+  }
+
+  /**
+   * @return The component to use as the button the user clicks to expand and view the
+   *    row edit modal for a particular row. Takes a single row prop and should emit a
+   *    'edit-modal' event when clicked with the row as the event.
+   */
+  getRowExpandButtonComponent() {
+    return GridViewRowExpandButton
   }
 
   getIconClass() {